Reorganizing model namespaces
