+++ ExcelReader/Deserialization/SharedStringsModels/SharedStrings.cs
@@ -1,0 +1,6 @@
+﻿namespace ExcelReader.Deserialization.SharedStringsModels
+{
+    class SharedStrings
+    {
+    }
+}
