+++ ExcelReader/Deserialization/SheetModels/Sheet.cs
@@ -1,0 +1,6 @@
+﻿namespace ExcelReader.Deserialization.SheetModels
+{
+    class Sheet
+    {
+    }
+}
