+++ ExcelReader/Deserialization/StylesModels/Styles.cs
@@ -1,0 +1,6 @@
+﻿namespace ExcelReader.Deserialization.StylesModels
+{
+    class Styles
+    {
+    }
+}
