+++ ExcelReader/Deserialization/WorkbookModels/Workbook.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Xml.Serialization;
 
-namespace ExcelReader.Models
+namespace ExcelReader.Deserialization.WorkbookModels
 {
     [Serializable, XmlRoot(ElementName = "workbook", Namespace = "http://schemas.openxmlformats.org/spreadsheetml/2006/main")]
     public class Workbook
