+++ ExcelReader/ExcelReader.csproj
@@ -11,4 +11,8 @@
     </AssemblyAttribute>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Models\Workbook\" />
+  </ItemGroup>
+
 </Project>
