--- ExcelReader/Models/SharedStrings.cs
@@ -1,6 +1,0 @@
-﻿namespace ExcelReader.Models
-{
-    class SharedStrings
-    {
-    }
-}
