--- ExcelReader/Models/Styles.cs
@@ -1,6 +1,0 @@
-﻿namespace ExcelReader.Models
-{
-    class Styles
-    {
-    }
-}
